requirements.txt
pandas==1.3.5
Pillow==9.0.0
streamlit==1.15.2
